+++ src/wrappers/mj_data.rs
@@ -153,7 +153,7 @@ impl<M: Deref<Target = MjModel>> MjData<M> {
             let qacc_smooth = nv_range;
             let qfrc_constraint = nv_range;
             let qfrc_inverse = nv_range;
-            
+
             let qfrc_spring = nv_range;
             let qfrc_damper = nv_range;
             let qfrc_gravcomp = nv_range;
@@ -892,6 +892,59 @@ impl<M: Deref<Target = MjModel>> MjData<M> {
         Ok((geom_id, dist))
     }
 
+    /// Intersect ray with visible flexes.
+    /// Return distance to nearest surface, or -1 if no intersection.
+    /// If `vertid` is `Some`, it will be filled with the id of the nearest vertex.
+    /// If `normal_out` is `Some`, it will be filled with the surface normal at the intersection.
+    /// `flex_layer`, `flg_vert`, `flg_edge`, `flg_face`, `flg_skin` and `flexid` control what and where to intersect.
+    pub fn ray_flex(
+        &self, flex_layer: i32, flg_vert: bool, flg_edge: bool, flg_face: bool, flg_skin: bool, flexid: i32,
+        pnt: &[MjtNum; 3], vec: &[MjtNum; 3],
+        vertid: Option<&mut i32>, normal_out: Option<&mut [MjtNum; 3]>
+    ) -> mjtNum {
+        unsafe { mj_rayFlex(
+            self.model.ffi(), self.ffi(),
+            flex_layer, flg_vert as MjtByte, flg_edge as MjtByte, flg_face as MjtByte, flg_skin as MjtByte, flexid,
+            pnt, vec,
+            vertid.map_or(ptr::null_mut(), |x| x), normal_out.map_or(ptr::null_mut(), |x| x)
+        ) }
+    }
+
+    /// Copies data state from `src` to `self` based on the specified `spec` combination of `mjtState` flags.
+    pub fn copy_state_from_data(&mut self, src: &MjData<M>, spec: u32) {
+        unsafe {
+            mj_copyState(self.model.ffi(), src.ffi(), self.ffi_mut(), spec as i32);
+        }
+    }
+
+    /// Intersect ray with hfield.
+    /// Returns the distance to the intersection, or -1.0 if no intersection.
+    pub fn ray_hfield(
+        &self, geom_id: i32, pnt: &[MjtNum; 3], vec: &[MjtNum; 3], normal_out: Option<&mut [MjtNum; 3]>
+    ) -> mjtNum {
+        unsafe {
+            mj_rayHfield(self.model.ffi(), self.ffi(), geom_id, pnt, vec, normal_out.map_or(ptr::null_mut(), |x| x))
+        }
+    }
+
+    /// Intersect ray with mesh.
+    /// Returns the distance to the intersection, or -1.0 if no intersection.
+    pub fn ray_mesh(
+        &self, geom_id: i32, pnt: &[MjtNum; 3], vec: &[MjtNum; 3], normal_out: Option<&mut [MjtNum; 3]>
+    ) -> mjtNum {
+        unsafe {
+            mj_rayMesh(self.model.ffi(), self.ffi(), geom_id, pnt, vec, normal_out.map_or(ptr::null_mut(), |x| x))
+        }
+    }
+
+    /// Apply Cartesian force and torque to a point on a body, and add the result to `qfrc_target`.
+    pub fn apply_ft(&mut self, force: &[MjtNum; 3], torque: &[MjtNum; 3], point: &[MjtNum; 3], body: i32, qfrc_target: &mut [MjtNum]) {
+        assert!(qfrc_target.len() >= self.model.ffi().nv as usize, "qfrc_target must be at least nv length");
+        unsafe {
+            mj_applyFT(self.model.ffi(), self.ffi_mut(), force, torque, point, body, qfrc_target.as_mut_ptr());
+        }
+    }
+
     /// Reads data's state into `destination`. The `spec` parameter is a bit mask of [`MjtState`] elements,
     /// which controls what state gets copied. The `destination` parameter is a mutable
     /// slice to the location into which the state will be written.
@@ -1338,7 +1391,7 @@ info_with_view!(
 // Camera view
 /**************************************************************************************************/
 info_with_view!(Data, camera, cam_, [xpos: MjtNum, xmat: MjtNum], [], M: Deref<Target = MjModel>);
-    
+
 /**************************************************************************************************/
 // Site view
 /**************************************************************************************************/
@@ -1366,6 +1419,10 @@ mod test {
 
     const MODEL: &str = "
 <mujoco>
+  <asset>
+    <mesh name=\"cube\" vertex=\"-0.5 -0.5 -0.5  0.5 -0.5 -0.5  -0.5  0.5 -0.5  0.5  0.5 -0.5  -0.5 -0.5  0.5  0.5 -0.5  0.5  -0.5  0.5  0.5  0.5  0.5  0.5\"/>
+    <hfield name=\"terrain\" nrow=\"10\" ncol=\"10\" size=\"10 10 .1 .1\"/>
+  </asset>
   <worldbody>
     <light ambient=\"0.2 0.2 0.2\"/>
     <body name=\"ball\" pos=\".2 .2 .1\">
@@ -1379,6 +1436,8 @@ mod test {
     </body>
 
     <geom name=\"floor1\" type=\"plane\" size=\"10 10 1\" solref=\"0.004 1.0\"/>
+    <geom name=\"mesh_cube\" type=\"mesh\" mesh=\"cube\" pos=\"2 2 0.5\"/>
+    <geom name=\"hfield_terrain\" type=\"hfield\" hfield=\"terrain\" pos=\"-2 -2 0\"/>
   </worldbody>
 </mujoco>";
 
@@ -1958,5 +2017,83 @@ mod test {
         let res = data.multi_ray(&pos, &ray_vecs, None, false, -1, 10.0, Some(&mut bad_normals));
         assert!(res.is_err());
         assert_eq!(res.unwrap_err().kind(), std::io::ErrorKind::InvalidInput);
+    }
+
+    #[test]
+    fn test_copy_state_from_data() {
+        let model = MjModel::from_xml_string(MODEL).unwrap();
+        let mut data1 = model.make_data();
+        let mut data2 = model.make_data();
+
+        data1.set_time(1.23);
+        data2.copy_state_from_data(&data1, MjtState::mjSTATE_TIME as u32);
+
+        assert_eq!(data2.time(), 1.23);
+    }
+
+    #[test]
+    fn test_ray_flex() {
+        let model = MjModel::from_xml_string(MODEL).unwrap();
+        let data = model.make_data();
+        let pos = [0.0; 3];
+        let vec = [1.0, 0.0, 0.0];
+
+        // Just checking that it runs without crashing, since we have no flex in MODEL
+        let dist = data.ray_flex(0, false, false, false, false, -1, &pos, &vec, None, None);
+        assert_eq!(dist, -1.0);
+    }
+
+    #[test]
+    fn test_ray_mesh_hfield() {
+        let model = MjModel::from_xml_string(MODEL).unwrap();
+        let mut data = model.make_data();
+
+        let mesh_id = model.geom("mesh_cube").unwrap().id as i32;
+        let hfield_id = model.geom("hfield_terrain").unwrap().id as i32;
+
+        data.forward_kinematics();
+
+        // Ray should hit mesh (centered at 2,2,0.5, size 1x1x1)
+        // Top surface of the mesh is at z=0.5 + 0.5 = 1.0. 
+        // Ray starts at z=5.0 and goes straight down [0, 0, -1].
+        // Intersection should be at exactly distance 4.0.
+        let mesh_dist = data.ray_mesh(mesh_id, &[2.0, 2.0, 5.0], &[0.0, 0.0, -1.0], None);
+        assert_relative_eq!(mesh_dist, 4.0, epsilon=1e-5);
+
+        // Ray should hit hfield (centered at -2,-2,0)
+        // Default hfield with no data acts as a plane at z=0.
+        // Intersection should be at exactly distance 5.0.
+        let hfield_dist = data.ray_hfield(hfield_id, &[-2.0, -2.0, 5.0], &[0.0, 0.0, -1.0], None);
+        assert_relative_eq!(hfield_dist, 5.0, epsilon=1e-5);
+    }
+
+    #[test]
+    fn test_apply_ft() {
+        let model = MjModel::from_xml_string(MODEL).unwrap();
+        let mut data = model.make_data();
+        let nv = model.nv() as usize;
+        let body_id = model.body("ball").unwrap().id as i32;
+        let mut qfrc = vec![0.0; nv];
+
+        data.forward();
+
+        // Apply force/torque at the ball's center of mass (pos = [0.2, 0.2, 0.1]) in global frame
+        let force = [1.5, 2.5, 3.5];
+        let torque = [0.1, 0.2, 0.3];
+        let point = [0.2, 0.2, 0.1]; 
+
+        data.apply_ft(&force, &torque, &point, body_id, &mut qfrc);
+
+        // The "ball" has a free joint.
+        // In MuJoCo, for a free joint, the first 3 DOFs are translation (linear),
+        // and the next 3 are the rotational DOFs.
+        // Since we applied it exactly at the COM, there should be no induced torque from the force position.
+        let dof_adr = model.body_dofadr()[body_id as usize] as usize;
+        assert!((qfrc[dof_adr] - 1.5).abs() < 1e-5);
+        assert!((qfrc[dof_adr + 1] - 2.5).abs() < 1e-5);
+        assert!((qfrc[dof_adr + 2] - 3.5).abs() < 1e-5);
+        assert!((qfrc[dof_adr + 3] - 0.1).abs() < 1e-5);
+        assert!((qfrc[dof_adr + 4] - 0.2).abs() < 1e-5);
+        assert!((qfrc[dof_adr + 5] - 0.3).abs() < 1e-5);
     }
 }